Making it possible to aggregate several requests in the Rack middleware
If using the Rack middleware to process a lot of requests, it can be quite
a pain to make sense of the sheer volume of files generated (one per request).

This patch adds a new `:max_requests` to the middleware, which will then
starts aggregating profiles across requests and will only dump the result file
when it has profiled `:max_requests` requests or when the current process shuts
down, whichever happens first.

Kept total backward compatibility.

Added tests on the new feature.