Better require_relative loading
Fixes to use DLEXT because e.g. Macs ".bundle" files and not ".so"

Also we wipe the ext dir in our installs since those files are normally
not used in the runtime at all.  The library is copied into the lib
dir though.

In case the ext dir was deliberate for development process that I'm
not aware of, I retained that as a final fallback.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>